requirements.txt
Django==2.2.7
django-geojson==2.12.0
django-leaflet==0.26.0
jsonfield==2.0.2
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
translitua==1.3.0
